Fix all unit tests
LSM Get checks in all levels if key is not found in memtable or level 0 to find latest version.
Fix parseItem. Breakage caused by how we store key in KVItem.
 Fix storing/reading readTs from badger head

+++ y/iterator.go
@@ -34,8 +34,8 @@ type ValueStruct struct {
 }
 
 // EncodedSize is the size of the ValueStruct when encoded
-func (v *ValueStruct) EncodedSize() int {
-	return len(v.Value) + 2
+func (v *ValueStruct) EncodedSize() uint16 {
+	return uint16(len(v.Value) + 2)
 }
 
 // Decode uses the length of the slice to infer the length of the Value field.
